+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Release Notes for Cloudflare Turnstile
 
+## 1.0.1 - 2023-04-13
+
+- Added `passes` and `fails` methods
+- Updated readme with new methods and on how to flag the message as spam
+
 ## 1.0.0 - 2023-01-06
 - Initial release